Moved CSV commands into unsync_csv.
