+++ unsync_csv/usync_csv/__init__.py
@@ -1,0 +1,9 @@
+__version__ = '0.0.1'
+
+unsync_prefix = 'csv'
+
+unsync_commands = [
+    'commands.export',
+    'commands.full_export',
+    'commands.csv_import',
+]
